updated sandcat
