+++ app/extensions/contact/gist.py
@@ -8,5 +8,5 @@ def load():
 class GIST(Extension):
 
     def __init__(self):
-        super().__init__(('gist.go', 'contact'))
+        super().__init__([('gist.go', 'contact')])
         self.dependencies = ['github.com/google/go-github/github', 'golang.org/x/oauth2']
